Calculate list price according to cost and utility values
Added a feature on Quotes modues to use or modify the cost of a product or service and specify an ammount or percentage of utility and calculate the value of List Price according to the cost and utility specified.
Add, fields for utility and utility type (amount or percentage) similar to discount and discount type fields.